server/requirements.txt
anyio==3.7.0
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
fastapi==0.99.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
pydantic==1.10.10
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.0
uvicorn==0.22.0
